use base class for MetaDisplayEvent
