+++ src/Event/TimesheetMetaDisplayEvent.php
@@ -9,15 +9,14 @@
 
 namespace App\Event;
 
-use App\Entity\MetaTableTypeInterface;
-use App\Repository\Query\BaseQuery;
 use App\Repository\Query\TimesheetQuery;
-use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * Dynamically find possible meta fields for a timesheet query.
+ *
+ * @method TimesheetQuery getQuery()
  */
-final class TimesheetMetaDisplayEvent extends Event implements MetaDisplayEventInterface
+final class TimesheetMetaDisplayEvent extends AbstractMetaDisplayEvent
 {
     public const EXPORT = 'export';
     public const TIMESHEET = 'timesheet';
@@ -25,55 +24,8 @@ final class TimesheetMetaDisplayEvent extends Event implements MetaDisplayEventI
     public const TIMESHEET_EXPORT = 'timesheet-export';
     public const TEAM_TIMESHEET_EXPORT = 'team-timesheet-export';
 
-    /**
-     * @var TimesheetQuery
-     */
-    private $query;
-    /**
-     * @var string
-     */
-    private $location;
-    /**
-     * @var MetaTableTypeInterface[]
-     */
-    private $fields = [];
-
     public function __construct(TimesheetQuery $query, string $location)
     {
-        $this->query = $query;
-        $this->location = $location;
-    }
-
-    /**
-     * If you want to filter where your meta-field will be displayed, use the query settings.
-     *
-     * @return TimesheetQuery
-     */
-    public function getQuery(): BaseQuery
-    {
-        return $this->query;
-    }
-
-    /**
-     * If you want to filter where your meta-field will be displayed, check the current location.
-     *
-     * @return string
-     */
-    public function getLocation(): string
-    {
-        return $this->location;
-    }
-
-    public function addField(MetaTableTypeInterface $meta)
-    {
-        $this->fields[] = $meta;
-    }
-
-    /**
-     * @return MetaTableTypeInterface[]
-     */
-    public function getFields(): array
-    {
-        return $this->fields;
+        parent::__construct($query, $location);
     }
 }
